chore( abstracted cookie interaction out of auth file )
