+++ lib/components/SimpleAuthForm.js
@@ -37,6 +37,8 @@ class SimpleAuthForm extends React.Component {
 
   async logOut() {
     // pending view state set here
+    console.log(this.props.store.token)
+
     const logOutRequest = await aposLogOut(this.props.store.token, this.props.store)
     // pending vew state resolved here
   }
