+++ lib/hoc/default.js
@@ -3,13 +3,13 @@
 import Styles from '../components/Styles'
 import Header from '../components/Header'
 import Footer from '../components/Footer'
-import {getValueFromServerCookie, getCookieValue} from '../utils/auth'
+import {getCookieValueFromServer, getCookieValue} from '../utils/cookie'
 
 
 
 export default Page => class DefaultPage extends React.Component {
   static async getInitialProps (ctx) {
-    const loggedUser = process.browser ? getCookieValue('user', true) : getValueFromServerCookie(ctx.req, 'user', true)
+    const loggedUser = process.browser ? getCookieValue('user', true) : getCookieValueFromServer(ctx.req, 'user', true)
     let pageProps = {}
     if (typeof Page.getInitialProps === 'function') {
       pageProps = await Page.getInitialProps.call(Page, ctx)
