+++ lib/with-mobx-store.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import {initializeStore} from '../store'
-import {getValueFromServerCookie} from './utils/auth'
+import {getCookieValueFromServer} from './utils/cookie'
 
 const isServer = typeof window === 'undefined'
 const getOrCreateStore = initialState => 
@@ -13,9 +13,9 @@ export default (App) => {
       // Get or Create the store with `undefined` as initialState
       // This allows you to set a custom default initialState
       const initialState = appContext.ctx.req ? {
-        user: getValueFromServerCookie(appContext.ctx.req, 'user', true),
-        loggedIn: getValueFromServerCookie(appContext.ctx.req, 'loggedIn'),
-        token: getValueFromServerCookie(appContext.ctx.req, 'token', true)
+        user: getCookieValueFromServer(appContext.ctx.req, 'user', true),
+        loggedIn: getCookieValueFromServer(appContext.ctx.req, 'loggedIn'),
+        token: getCookieValueFromServer(appContext.ctx.req, 'token', true)
       } : {}
       const mobxStore = getOrCreateStore(initialState)
